OrcLib: MftRecordAttribute: LoadBitField: fix bitmap use
The boost::dynamic_bitset class is tricky to use. The method append will
pad the data if smaller than block_type (the T type its built on).

BitmapAttribute::block_type was size_t while 'append' was called with
ULONG leading to corruption of the bitmap. Use now uint8_t as block_type
and added an helper to initialize byte to byte from a container. But be
careful to container inner type and endianness.

+++ src/OrcLib/MftRecordAttribute.h
@@ -269,7 +269,8 @@ class BitmapAttribute : public MftRecordAttribute
     friend class MFTRecord;
 
 public:
-    using block_type = size_t;
+    // BEWARE: boost::dynamic_bitset<T>::append will complete the missing bytes if type is smaller
+    using block_type = uint8_t;
 
     BitmapAttribute(PATTRIBUTE_RECORD_HEADER pHeader, MFTRecord* pRecord)
         : MftRecordAttribute(pHeader, pRecord) {
